refactor(tools): unify Crossref commands
