+++ docs/supabase/ADD_PRESENTER_SEARCH_REMOTE.sql
@@ -42,7 +42,7 @@ VALUES (
   'tool_use/search.yaml',
   ARRAY['researcher'],
   'toolUse',
-  ARRAY['bash', 'read_file', 'glob', 'grep', 'extract_figures', 'extract_bib_entries', 'extract_tikz_figures', 'arxiv_search', 'arxiv_metadata', 'download_arxiv_source', 'crossref_search', 'crossref_doi', 'zotero_add', 'zotero_search', 'zotero_export', 'web_search', 'web_fetch']
+  ARRAY['bash', 'read_file', 'glob', 'grep', 'extract_figures', 'extract_bib_entries', 'extract_tikz_figures', 'arxiv_search', 'arxiv_metadata', 'download_arxiv_source', 'crossref_search', 'zotero_add', 'zotero_search', 'zotero_export', 'web_search', 'web_fetch']
 );
 
 INSERT INTO remote_agents (name, description, storage_path, visibility, agent_category, tools)
