+++ src/tools/citation/CrossrefSearchTool.ts
@@ -2,6 +2,7 @@
 import {
   SortOrder,
   type QueryWorksParams,
+  type Work,
   type WorkSortOptions,
 } from '@jamesgopsill/crossref-client';
 import { z } from 'zod';
@@ -16,29 +17,40 @@ import { pluralize } from '@utils/text/stringUtils';
 import { CROSSREF_CONSTANTS, CrossrefClient } from './constants';
 import { rateLimitedRequest } from './rateLimiter';
 
-const CrossrefSearchInputSchema = z.strictObject({
-  query: z.string().describe('Bibliographic search query for Crossref works.'),
-  rows: z
-    .int()
-    .positive()
-    .max(CROSSREF_CONSTANTS.MAX_ROWS)
-    .nullish()
-    .transform((v) => v ?? CROSSREF_CONSTANTS.DEFAULT_ROWS)
-    .describe('Maximum number of works to return.'),
-  offset: z
-    .int()
-    .min(0)
-    .nullish()
-    .describe('Zero-based result offset for pagination.'),
-  sort: z.string().nullish().describe('Crossref sort field to apply.'),
-  order: z.enum(['asc', 'desc']).nullish().describe('Crossref sort order.'),
-  // Filter as Crossref filter string format (e.g., "from-pub-date:2023,has-orcid:true")
-  // Object format removed due to OpenAI JSON Schema limitations with z.record()
-  filter: z
-    .string()
-    .nullish()
-    .describe('Crossref filter string, e.g. "from-pub-date:2023".'),
-});
+const CrossrefSearchInputSchema = z.discriminatedUnion('command', [
+  z.strictObject({
+    command: z.literal('search').describe('Search Crossref works.'),
+    query: z
+      .string()
+      .describe('Bibliographic search query for Crossref works.'),
+    rows: z
+      .int()
+      .positive()
+      .max(CROSSREF_CONSTANTS.MAX_ROWS)
+      .nullish()
+      .transform((v) => v ?? CROSSREF_CONSTANTS.DEFAULT_ROWS)
+      .describe('Maximum number of works to return.'),
+    offset: z
+      .int()
+      .min(0)
+      .nullish()
+      .describe('Zero-based result offset for pagination.'),
+    sort: z.string().nullish().describe('Crossref sort field to apply.'),
+    order: z.enum(['asc', 'desc']).nullish().describe('Crossref sort order.'),
+    // Filter as Crossref filter string format (e.g., "from-pub-date:2023,has-orcid:true")
+    // Object format removed due to OpenAI JSON Schema limitations with z.record()
+    filter: z
+      .string()
+      .nullish()
+      .describe('Crossref filter string, e.g. "from-pub-date:2023".'),
+  }),
+  z.strictObject({
+    command: z.literal('doi').describe('Look up one DOI in Crossref.'),
+    doi: z
+      .string()
+      .describe('DOI to look up, with or without a DOI URL prefix.'),
+  }),
+]);
 
 export type CrossrefSearchInput = z.infer<typeof CrossrefSearchInputSchema>;
 
@@ -53,10 +65,53 @@ type ExtendedQueryWorksParams = QueryWorksParams & { filter?: string };
 export class CrossrefSearchTool extends defineTool({
   name: 'crossref_search',
   parallelSafe: true,
-  description: 'Search Crossref works and return top matches.',
+  description:
+    'Search Crossref works or look up detailed metadata for a DOI. Use command="search" with query, or command="doi" with doi.',
   schema: CrossrefSearchInputSchema,
 }) {
   protected async execute(input: CrossrefSearchInput): Promise<ToolResult> {
+    if (input.command === 'doi') {
+      const trimmedDoi = requireNonEmptyString(input.doi, 'DOI');
+
+      const response = await wrapApiCall(
+        () =>
+          rateLimitedRequest(
+            'crossref',
+            CROSSREF_CONSTANTS.RATE_LIMIT_DELAY_MS,
+            'Crossref DOI lookup',
+            () => CrossrefClient.work(trimmedDoi),
+          ),
+        'Crossref lookup failed',
+      );
+
+      if (!response.ok || !response.content?.message) {
+        throw new ToolError('Crossref response did not include metadata.');
+      }
+
+      const work: Work = response.content.message;
+      const metadata = {
+        doi: work.DOI,
+        title: work.title?.[0] ?? null,
+        titles: work.title ?? [],
+        publisher: work.publisher,
+        type: work.type,
+        abstract: work.abstract ?? null,
+        description: null, // Not exposed by the library type.
+        created: work.created,
+        published: work.published ?? null,
+        url: work.resource?.primary?.URL ?? null,
+        language: work.language ?? null,
+        authors: work.author ?? [],
+        licenses: work.license ?? [],
+      };
+
+      return {
+        status: 'executed',
+        summary: `Retrieved: DOI ${metadata.doi}`,
+        output: JSON.stringify(metadata, null, 2),
+      };
+    }
+
     const trimmedQuery = requireNonEmptyString(input.query, 'Search query');
 
     const options: ExtendedQueryWorksParams = {
